EncryptionMethods/src/main/java/utils/UtilFunctions.java
package utils;

import java.math.BigInteger;
import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;

/**
 * Created by Алексей on 24.12.2016.
 */
public abstract class UtilFunctions {
    private static MessageDigest messageDigest;

    public static String md5Custom(String st) {
        if (isNullString(st)) {
            return "";
        }

        byte[] digest = new byte[0];
        try {
            messageDigest = MessageDigest.getInstance("MD5");
            messageDigest.reset();
            messageDigest.update(st.getBytes());
            digest = messageDigest.digest();
        } catch (NoSuchAlgorithmException e) {
            // тут можно обработать ошибку
            // возникает она если в передаваемый алгоритм в getInstance(,,,) не существует
            e.printStackTrace();
        }

        BigInteger bigInt = new BigInteger(1, digest);
        String md5Hex = bigInt.toString(16);

        while (md5Hex.length() < 32) {
            md5Hex = "0" + md5Hex;
        }

        return md5Hex;
    }

    public static boolean isNullString(String value) {
        return value == null || value.trim().isEmpty();
    }
}
